Remove Capability and move destination related methods to another trait
